Assembler Release, corrections
Assembler Release: This feature is EXTREMELY experimental, to be honest, this code is so bad that I should probably be tried for making it - but it does its job. It is extremely basic, so no ORG or width directives on arguments, instead declarations and looking at the arguments themselves. I will document the ASM Dialect on the Wiki tomorrow, as it is late. Its implementation into Main was also a quick job to ensure that the assembler was on the repository by the end of the night, it is 00:08AM GMT as of writing.\n SYS and ARH Corrections: Small Corrections, mainly removal of passed arguments that served no purpose in the instructions they were being passed to.